requirements/requirements-dev.txt
# TensorFlow dengan GPU support
tensorflow[and-cuda]

# Keras (standalone)
keras

# Image processing
Pillow
opencv-python

# Data manipulation
numpy
pandas

# Visualization
matplotlib
seaborn

# Model evaluation
scikit-learn

# Progress bar
tqdm

# Utilities
pyyaml